[HUDI-7742] Move Hadoop-dependent reader util classes to hudi-hadoop-common module
